requirements.txt
# STALWART - Core Dependencies
# Version: 1.0.0
# Date: 2026-02-16

# Core scientific computing
numpy>=1.24.0
scipy>=1.11.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Machine learning
scikit-learn>=1.3.0
xgboost>=1.7.0
lightgbm>=3.3.0
optuna>=3.0.0

# Deep learning (optional)
# tensorflow>=2.13.0
# torch>=2.0.0

# Signal processing
pywt>=1.4.0
rainflow>=3.0.0
control>=0.9.0

# Data validation
pydantic>=2.0.0

# Sensor communication
pyserial>=3.5
smbus2>=0.4.0
# RPi.GPIO>=0.7.0  # Raspberry Pi only
# Adafruit-Blinka>=8.0.0

# Database
psycopg2>=2.9.0
sqlalchemy>=2.0.0
timescaledb>=2.8.0
influxdb>=5.3.0
redis>=4.5.0

# Web API
fastapi>=0.95.0
uvicorn>=0.21.0
python-multipart>=0.0.6
httpx>=0.24.0

# Dashboard
plotly>=5.13.0
dash>=2.9.0
streamlit>=1.20.0
grafana-api>=1.0.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
python-dotenv>=1.0.0
requests>=2.31.0
click>=8.1.0
rich>=13.0.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-asyncio>=0.21.0
pytest-mock>=3.10.0

# Code quality
black>=22.0.0
flake8>=4.0.0
mypy>=0.9.0
pre-commit>=2.0.0
isort>=5.12.0

# Documentation
sphinx>=4.0.0
sphinx-rtd-theme>=1.0.0
sphinx-autodoc-typehints>=1.19.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.0.0
ipywidgets>=8.0.0
